Vgff screens

+++ src/screens/EditStation/index.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { HeaderApp } from "../../components/HeaderApp";
-import { Container, ContainerButtons, Image, ImageContainer, ItemContainer, ListContainer, PartnerContainer, PartnerHeader, SensorPartnerContainer, TitleItem, TitlePartnerSensorContainer } from "./styles";
+import { ButtonPublicPrivate, ButtonText, Container, ContainerButtons, ContainerPublicPrivate, Image, ImageContainer, ItemContainer, ListContainer, PartnerContainer, PartnerHeader, SensorPartnerContainer, TitleItem, TitlePartnerSensorContainer } from "./styles";
 import { WeatherStationsService } from "../../services/WeatherStationService";
 import { useNavigation } from "@react-navigation/native";
 import { ListEmpty } from "../../components/ListEmpty";
@@ -18,6 +18,7 @@ import { Button } from "../../components/Button";
 import { ModalImagePicker } from "../../components/ModalImagePicker";
 import MantainerData from "../../interfaces/weatherStation/MantainerData";
 import { ModalMantainer } from "../../components/ModalMantainer";
+import Toast from 'react-native-toast-message'
 
 type Props = {
     stationId?: string;
@@ -32,16 +33,18 @@ export function EditStation({ stationId } : Props){
             longitude: '',
             altitudeMSL: '',
             partners: [],
-            image: '',
+            photoBase64: '',
             sensors: [],
         }
     );
     const [mantainer, setMantainer] = useState<MantainerData[]>();
     const [originalWeatherStation, setOriginalWeatherStation] = useState<WeatherStationData>();
+    const [weatherStationPhoto, setWeatherStationPhoto] = useState<String>();
     const [showModalPartner, setShowModalPartner] = useState(false);
     const [showModalImage, setShowModalImage] = useState(false);
     const [showModalMantainer, setShowModalMantainer] = useState(false);
     const [partner, setPartner] = useState()
+    const [loading, setLoading] = useState<boolean>(false);
     const service = new WeatherStationsService();
     const navigate = useNavigation<StackType>();
 
@@ -54,6 +57,18 @@ export function EditStation({ stationId } : Props){
     async function getMaintainers(){
         const response = await service.getAllMaintainersByWeatherStationId(stationId || '1')
         setMantainer(response.data)
+    }
+
+    async function getWeatherStationPhoto(stationid : string){
+        try {
+            const response = await service.getWeatherStationPhoto(stationid)
+            if(response){
+                setWeatherStationPhoto(response)
+            }
+        } catch (error) {
+            
+        }
+        
     }
 
     function handleBack(){
@@ -111,6 +126,7 @@ export function EditStation({ stationId } : Props){
     }
 
     async function UpdateStation() {
+        setLoading(true)
             try {
                 const response = await service.updateWeatherStation(weatherStation);
                 if(response){
@@ -120,6 +136,7 @@ export function EditStation({ stationId } : Props){
             } catch (error) {
                 console.log(error)
             }
+        setLoading(false)
     }
 
     function hasStationChanged(): boolean {
@@ -143,16 +160,26 @@ export function EditStation({ stationId } : Props){
         
     }
 
-    async function handleDeleteMantainer(mantainerId : number){
+    async function handleDeleteMantainer(mantainerId : string){
         try {
-            const response = await service.deleteMantainer(weatherStation.id, mantainerId);
-            if(response){
-                alert("Mantenedor removido com sucesso")
-                getMaintainers()
-                getStation()
+            if(mantainer?.length == 1){
+                Toast.show({
+                    type: 'error',
+                    text1: 'Erro ao remover mantenedor',
+                    text2: 'A estação deve ter no mínimo um mantenedor',
+                    position: 'bottom',
+                    bottomOffset: 60
+                })
+            } else {
+                const response = await service.deleteMantainer(weatherStation.id, mantainerId);
+                if(response){
+                    alert("Mantenedor removido com sucesso")
+                    getMaintainers()
+                    getStation()
+                }
             }
         } catch (error) {
-            alert("Erro ao remover mantenedor")
+            // alert("Erro ao remover mantenedor")
         }
     }
     
@@ -160,6 +187,7 @@ export function EditStation({ stationId } : Props){
     useEffect(() => {
         getStation()
         getMaintainers()
+        getWeatherStationPhoto(stationId || '0')
     }, [])
 
     return(
@@ -245,17 +273,28 @@ export function EditStation({ stationId } : Props){
                         )
                     }
                 </ItemContainer>
-
                 <ItemContainer>
                     <TitleItem>Foto</TitleItem>
                     <ImageContainer>
-                        <Image source={weatherStation?.image ? {uri: weatherStation.image} : require('../../assets/aty.png')} />
+                        <Image source={weatherStationPhoto != undefined ? { uri: `data:image/jpeg;base64,${weatherStationPhoto}` } : require('../../assets/aty.png')}/>
                     </ImageContainer>
                 </ItemContainer>
 
+                {/* <PartnerHeader>
+                    <TitlePartnerSensorContainer>Tipo da Estação:</TitlePartnerSensorContainer>
+                </PartnerHeader>
+                <ContainerPublicPrivate>
+                    <ButtonPublicPrivate onPress={() => setWeatherStation(prevState => ({ ...prevState, isPrivate: false }))} bgColor={weatherStation.isPrivate ? "BLUE" : "GREEN"}>
+                        <ButtonText>Estação Pública</ButtonText>    
+                    </ButtonPublicPrivate>
+                    <ButtonPublicPrivate onPress={() => setWeatherStation(prevState => ({ ...prevState, isPrivate: true }))} bgColor={weatherStation.isPrivate ? "GREEN" : "BLUE"}>
+                        <ButtonText>Estação Privada</ButtonText>    
+                    </ButtonPublicPrivate>
+                </ContainerPublicPrivate> */}
+
                 <ContainerButtons>
-                        <Button title={weatherStation?.image ? "Adicionar nova foto" : "Adicionar Foto"} onPress={() => setShowModalImage(true)} color="SECONDARY" />
-                        <Button title="Atualizar Estação" onPress={() =>{hasStationChanged() ? UpdateStation() : console.log("Não foi")} } color="PRIMARY" />
+                        <Button title={weatherStationPhoto ? "Adicionar nova foto" : "Adicionar Foto"} onPress={() => setShowModalImage(true)} color="SECONDARY" />
+                        <Button title="Atualizar Estação" onPress={() =>{hasStationChanged() ? UpdateStation() : console.log("Não foi")} } color="PRIMARY" isLoading={loading} />
                 </ContainerButtons>
             </ListContainer>
 
@@ -273,8 +312,9 @@ export function EditStation({ stationId } : Props){
                     onSubmit={(data) => {
                         setWeatherStation((prev: any) => ({
                             ...prev,
-                            image: data
+                            photoBase64: data
                         }));
+                        setWeatherStationPhoto(data)
                         setShowModalImage(false)
                     }}
                 />
